kms: Fail post_update if shutdown has commenced
This is necessary with triple buffering because flush_callbacks can
result in new callbacks being queued, and we would fail the assertion
in meta_kms_impl_device_atomic_finalize on shutdown.

It's a harmless failure, but still not a good look to SIGABRT. So avoid
the problem by not queuing new updates when shutdown has already begun.