clutter/frame-clock: Target the correct next_presentation_time_us
Where "correct" means a different time to when we expect the already
dispatched update to present.

This logic error has existed for a long time. It didn't affect double
buffering in any noticeable way, probably because the "early presented
event" logic corrected it.

But for triple buffering this matters a lot more. Otherwise we get
stuck in triple buffering for too long by always scheduling the next
update earlier than necessary.